Add login form error message + Update profile page
Added functionality to display an error message on the login form when the credentials entered are incorrect.

Updated the profile page to display user information.

Removed the use of local storage for the user.

Removed the header component and integrated its functionality into the profile component.

Added the ability to modify user information on the profile page.